Update USB HID patch to work on MicroPython v1.20.
TinyUSB support was refactored since v1.17 so this has been rewritten, but
is entirely based on https://github.com/noobee/micropython/tree/usb-hid

Updated to use a patch file rather than completely overwriting, this should
help make it keep working against future versions.

Because the way the USB names are handled is different now, apply that in
a separate patch.

Tested against the nightly, but this should work on the v1.20 tag.

Signed-off-by: Jim Mussared <jim.mussared@gmail.com>